gsm.mapping: Good Statistical Monitoring Mapping

The Good Statistical Monitoring or 'gsm' R package provides a framework for statistical data monitoring. 'gsm.mapping' provides workflows to apply the necessary data transformation from raw/source datasets to appropriate domains.

See also

Author

Maintainer: Jeremy Wildfire jwildfire@gmail.com

Authors:

Laura Maxwell laura.maxwell@atorusresearch.com

Zelos Zhu zelos.zhu@atorusresearch.com

Other contributors:

Gilead Sciences [copyright holder]